requirements.txt
arxiv
requests
tqdm
pdfminer.six
markdownify
Pillow
openai
weaviate-client
python-dotenv
PyYAML
